[250907] Day24 - 과제 제출
