+++ my-app/my-app/src/app/boards/BoardList.module.css
@@ -72,18 +72,25 @@
 .FetchBoard {
   max-width: 74rem;
   width: 100%;
-  height: 2.75rem;
-
-  display: flex;
-  padding: 0.75rem 1.5rem;
-  align-items: center;
-  gap: 0.5rem;
-  align-self: stretch;
+  height: 34.3rem;
   /* 
   display: -webkit-box;
   -webkit-box-orient: vertical;
   -webkit-line-clamp: 1;
   flex: 1 0 0; */
+  display: grid;
+}
+
+.FetchBoard_list {
+  display: flex;
+  width: 74rem;
+  /* width: 100%; */
+  height: 2.75rem;
+
+  padding: 0.75rem 1.5rem;
+  align-items: center;
+  gap: 0.5rem;
+  align-self: stretch;
 
   border-radius: 0.5rem;
   border: 1px solid #f2f2f2;
