+++ my-app/my-app/src/components/boards-write/index.module.css
@@ -84,6 +84,27 @@
   background: var(--Gray-W, #fff);
 }
 
+.게시물등록_수정플레이스홀더 {
+  align-self: stretch;
+  max-width: 80rem;
+  width: 100%;
+  height: 3rem;
+
+  display: flex;
+  padding: 0.75rem 1rem;
+  flex-direction: column;
+  align-items: flex-start;
+  flex: 1 0 0;
+
+  color: var(--Gray-Gray-400, #ababab);
+  font-size: 1rem;
+  font-weight: 400;
+  line-height: 1.5rem; /* 150% */
+
+  border-radius: 0.5rem;
+  background: var(--gray-50, #f2f2f2);
+}
+
 .게시물등록_작성자and비밀번호 {
   display: flex;
   align-items: flex-start;
